`fn obmc_lowest_px`: Combined some `as` casts.

+++ src/decode.rs
@@ -4240,12 +4240,12 @@ unsafe fn obmc_lowest_px(
             if a_r.r#ref.r#ref[0] as libc::c_int > 0 {
                 let oh4 = imin(b_dim[1] as libc::c_int, 16) >> 1;
                 mc_lowest_px(
-                    &mut dst[(a_r.r#ref.r#ref[0] as libc::c_int - 1) as usize][is_chroma as usize],
+                    &mut dst[a_r.r#ref.r#ref[0] as usize - 1][is_chroma as usize],
                     t.by,
                     oh4 * 3 + 3 >> 2,
                     a_r.mv.mv[0].y as libc::c_int,
                     ss_ver,
-                    &(f.svc[(a_r.r#ref.r#ref[0] as libc::c_int - 1) as usize])[1],
+                    &(f.svc[a_r.r#ref.r#ref[0] as usize - 1])[1],
                 );
                 i += 1;
             }
@@ -4263,12 +4263,12 @@ unsafe fn obmc_lowest_px(
             if l_r.r#ref.r#ref[0] as libc::c_int > 0 {
                 let oh4_0 = iclip(l_b_dim[1] as libc::c_int, 2, b_dim[1] as libc::c_int);
                 mc_lowest_px(
-                    &mut dst[(l_r.r#ref.r#ref[0] as libc::c_int - 1) as usize][is_chroma as usize],
+                    &mut dst[l_r.r#ref.r#ref[0] as usize - 1][is_chroma as usize],
                     t.by + y,
                     oh4_0,
                     l_r.mv.mv[0].y as libc::c_int,
                     ss_ver,
-                    &f.svc[(l_r.r#ref.r#ref[0] as libc::c_int - 1) as usize][1],
+                    &f.svc[l_r.r#ref.r#ref[0] as usize - 1][1],
                 );
                 i_0 += 1;
             }
